requirements.txt
djangorestframework
mysqlclient
djangorestframework_simplejwt
django-environ
